Skip registered data files that don't exist
Not all data files are present for all years. If the data file
is validated (by having a key in the mapping), it can be skipped
if it doesn't exist.

The `read_data` function has to be split up so that a `KeyError`
is thrown immediately when a data file isn't found, rather then
when the generator is consumed.