Add run tracking
Add a table to track runs, and corresponding `Run` object.
Create a worker subclass that provides job tracking and
integrate it with the enqueue endpoint.
Add endpoint to get a run's status.
Incorporate a database connectivity check as part of the
app's health check.
Swap isolation level to read-committed, as read-repeatable is
not necessary with one worker and causes visibility errors
for the run status endpoint.
Implement retry on database connection so that clients don't
automatically fail during database startup.