Speed up bulk writes
Disable WAL archival and use `create table as` where possible in
order to speed up bulk writes (see [section
14.4.7](https://www.postgresql.org/docs/11/populate.html)). Because
the data can be reloaded if lost, it is not necessary to keep a full
logical WAL.

`vacuum analyze` is also explicitly run between staging and mart
load jobs, to ensure table statistics are up to date before the mart
load, speeding it up drastically. Autovacuum may not have time to
run in between jobs, so triggering a vacuum ensures it happens.